# artifacts/api-server/src/routes/auth.ts
import { Router } from "express";
import { db, usersTable, subscriptionsTable, knowledgeBaseTable, whatsappSessionsTable } from "@workspace/db";
import { eq } from "drizzle-orm";
import { RegisterBody, LoginBody } from "@workspace/api-zod";
import { hashPassword, generateToken, storeToken, getUserIdFromToken, removeToken, requireAuth, type AuthRequest } from "../lib/auth";

const router = Router();

router.post("/auth/register", async (req, res) => {
  const parsed = RegisterBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: "Validation error", message: parsed.error.message });
    return;
  }
  const { name, email, phone, businessName, password } = parsed.data;

  const existing = await db.select().from(usersTable).where(eq(usersTable.email, email));
  if (existing.length > 0) {
    res.status(400).json({ error: "User already exists", message: "Email is already registered" });
    return;
  }

  const [user] = await db.insert(usersTable).values({
    name,
    email,
    phone,
    businessName,
    passwordHash: hashPassword(password),
  }).returning();

  // Create trial subscription (2 days)
  const trialEnd = new Date();
  trialEnd.setDate(trialEnd.getDate() + 2);
  await db.insert(subscriptionsTable).values({
    userId: user.id,
    plan: "TRIAL",
    status: "TRIAL",
    endDate: trialEnd,
    trialEndDate: trialEnd,
    messagesLimit: 100,
  });

  // Create empty knowledge base
  await db.insert(knowledgeBaseTable).values({
    userId: user.id,
    tone: "friendly",
    personality: "helpful",
  });

  // Create WhatsApp session record
  await db.insert(whatsappSessionsTable).values({
    userId: user.id,
    status: "disconnected",
  });

  const token = generateToken(user.id);
  storeToken(token, user.id);

  res.status(201).json({
    user: {
      id: user.id,
      name: user.name,
      email: user.email,
      phone: user.phone,
      businessName: user.businessName,
      businessType: user.businessType,
      role: user.role,
      language: user.language,
      onboardingStep: user.onboardingStep,
      isVerified: user.isVerified,
      createdAt: user.createdAt,
    },
    token,
  });
});

router.post("/auth/login", async (req, res) => {
  const parsed = LoginBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: "Validation error", message: parsed.error.message });
    return;
  }
  const { email, password } = parsed.data;

  const [user] = await db.select().from(usersTable).where(eq(usersTable.email, email));
  if (!user || user.passwordHash !== hashPassword(password)) {
    res.status(401).json({ error: "Unauthorized", message: "Invalid email or password" });
    return;
  }

  const token = generateToken(user.id);
  storeToken(token, user.id);

  res.json({
    user: {
      id: user.id,
      name: user.name,
      email: user.email,
      phone: user.phone,
      businessName: user.businessName,
      businessType: user.businessType,
      role: user.role,
      language: user.language,
      onboardingStep: user.onboardingStep,
      isVerified: user.isVerified,
      createdAt: user.createdAt,
    },
    token,
  });
});

router.post("/auth/logout", requireAuth, (req, res) => {
  const auth = req.headers.authorization;
  const token = auth?.startsWith("Bearer ") ? auth.slice(7) : req.cookies?.token;
  if (token) removeToken(token);
  res.json({ success: true });
});

router.get("/auth/me", requireAuth, async (req, res) => {
  const user = (req as AuthRequest).user;
  res.json({
    id: user.id,
    name: user.name,
    email: user.email,
    phone: user.phone,
    businessName: user.businessName,
    businessType: user.businessType,
    role: user.role,
    language: user.language,
    onboardingStep: user.onboardingStep,
    isVerified: user.isVerified,
    createdAt: user.createdAt,
  });
});

export default router;