# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WhatsBiz AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ===================== BUSINESS =====================
  /business/profile:
    get:
      operationId: getBusinessProfile
      tags: [business]
      summary: Get business profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProfile"
    put:
      operationId: updateBusinessProfile
      tags: [business]
      summary: Update business profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBusinessProfileBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProfile"
  /business/onboarding:
    post:
      operationId: completeOnboarding
      tags: [business]
      summary: Complete onboarding step
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingResponse"

  # ===================== KNOWLEDGE BASE =====================
...
  /broadcasts:
    get:
      operationId: listBroadcasts
      tags: [broadcasts]
      summary: List broadcast campaigns
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BroadcastListResponse"
    post:
      operationId: createBroadcast
      tags: [broadcasts]
      summary: Create a broadcast
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBroadcastBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Broadcast"
  /broadcasts/{id}/send:
    post:
      operationId: sendBroadcast
      tags: [broadcasts]
      summary: Send / schedule a broadcast
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Broadcast"

  # ===================== ANALYTICS =====================
...
  /billing/initiate-payment:
    post:
      operationId: initiatePayment
      tags: [billing]
      summary: Initiate UPI payment for a plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiatePaymentBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentInitResponse"
  /billing/verify-payment:
    post:
      operationId: verifyPayment
      tags: [billing]
      summary: Submit UTR for manual payment verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPaymentBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]

    # ===================== AUTH =====================
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        businessName:
          type: string
        password:
          type: string
          minLength: 6
      required: [name, email, phone, businessName, password]
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        businessName:
          type: string
        businessType:
          type: string
          nullable: true
        role:
          type: string
          enum: [CLIENT, ADMIN, SUPPORT]
        language:
          type: string
        onboardingStep:
          type: integer
        isVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, phone, businessName, role, language, onboardingStep, isVerified, createdAt]

    # ===================== BUSINESS =====================
    BusinessProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        businessName:
          type: string
        businessType:
          type: string
          nullable: true
        businessSize:
          type: string
          nullable: true
        language:
          type: string
        onboardingStep:
          type: integer
        onboardingComplete:
          type: boolean
      required: [id, name, email, phone, businessName, language, onboardingStep, onboardingComplete]
    UpdateBusinessProfileBody:
      type: object
      properties:
        name:
          type: string
        businessName:
          type: string
        businessType:
          type: string
        businessSize:
          type: string
        language:
          type: string
    OnboardingBody:
      type: object
      properties:
        step:
          type: integer
        data:
          type: object
          additionalProperties: true
      required: [step, data]
    OnboardingResponse:
      type: object
      properties:
        nextStep:
          type: integer
        complete:
          type: boolean
      required: [nextStep, complete]

    # ===================== KNOWLEDGE BASE =====================
...
    Broadcast:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        message:
          type: string
        recipientType:
          type: string
          enum: [all, tag, custom]
        recipientCount:
          type: integer
        status:
          type: string
          enum: [draft, scheduled, sending, sent, failed]
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        sentAt:
          type: string
          format: date-time
          nullable: true
        deliveredCount:
          type: integer
        readCount:
          type: integer
        repliedCount:
          type: integer
        failedCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, message, recipientType, recipientCount, status, deliveredCount, readCount, repliedCount, failedCount, createdAt]
    BroadcastListResponse:
      type: object
      properties:
        broadcasts:
          type: array
          items:
            $ref: "#/components/schemas/Broadcast"
        total:
          type: integer
      required: [broadcasts, total]
    CreateBroadcastBody:
      type: object
      properties:
        name:
          type: string
        message:
          type: string
        recipientType:
          type: string
          enum: [all, tag, custom]
        recipients:
          type: array
          items:
            type: string
        scheduledAt:
          type: string
          format: date-time
      required: [name, message, recipientType]

    # ===================== ANALYTICS =====================
...
    InitiatePaymentBody:
      type: object
      properties:
        plan:
          type: string
          enum: [STARTER, PRO, BUSINESS]
      required: [plan]
    PaymentInitResponse:
      type: object
      properties:
        paymentId:
          type: string
        amount:
          type: integer
        upiId:
          type: string
        txnNote:
          type: string
        qrCodeUrl:
          type: string
          nullable: true
        upiDeepLinks:
          type: object
          properties:
            gpay:
              type: string
            phonepe:
              type: string
            paytm:
              type: string
          required: [gpay, phonepe, paytm]
      required: [paymentId, amount, upiId, txnNote, upiDeepLinks]
    VerifyPaymentBody:
      type: object
      properties:
        paymentId:
          type: string
        utr:
          type: string
      required: [paymentId, utr]
    Payment:
      type: object
      properties:
        id:
          type: string
        amount:
          type: integer
        plan:
          type: string
        status:
          type: string
          enum: [PENDING, AWAITING_VERIFICATION, VERIFIED, COMPLETED, FAILED, REFUNDED]
        utr:
          type: string
          nullable: true
        txnNote:
          type: string
        paidAt:
          type: string
          format: date-time
          nullable: true
        invoiceNumber:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, amount, plan, status, txnNote, createdAt]
...