# lib/api-zod/src/generated/types/broadcastListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsBiz AI API
 * OpenAPI spec version: 0.1.0
 */
import type { Broadcast } from "./broadcast";

export interface BroadcastListResponse {
  broadcasts: Broadcast[];
  total: number;
}