# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsBiz AI API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: string;
  name: string;
  email: string;
  phone: string;
  businessName: string;
  businessType?: string | null;
  role: UserRole;
  language: string;
  onboardingStep: number;
  isVerified: boolean;
  createdAt: Date;
}